@@ -39,47 +39,51 @@
     "vue": "3.5.13",
     "vue-eslint-parser": "9.4.3",
     "vue-router": "4.5.0",
-    "vue-tsc": "2.2.0",
-    "vue-types": "5.1.3"
+    "vue-tsc": "2.2.4",
+    "vue-types": "6.0.0"
   },
   "devDependencies": {
-    "@iconify/json": "2.2.292",
-    "@iconify/utils": "2.2.1",
-    "@types/node": "22.10.5",
-    "@typescript-eslint/eslint-plugin": "8.19.1",
-    "@typescript-eslint/parser": "8.19.1",
-    "@unocss/eslint-config": "0.65.3",
+    "@babel/eslint-parser": "^7.26.8",
+    "@iconify/json": "2.2.310",
+    "@iconify/utils": "2.3.0",
+    "@stylistic/stylelint-plugin": "^3.1.2",
+    "@types/node": "22.13.5",
+    "@typescript-eslint/eslint-plugin": "8.24.1",
+    "@typescript-eslint/parser": "8.24.1",
+    "@unocss/eslint-config": "66.0.0",
     "@vue/compiler-sfc": "3.5.13",
     "@vue/runtime-core": "3.5.13",
     "@vue/shared": "3.5.13",
     "autoprefixer": "10.4.20",
     "cross-env": "7.0.3",
-    "eslint": "8.57.0",
-    "eslint-config-prettier": "9.1.0",
-    "eslint-plugin-prettier": "5.2.1",
+    "eslint": "9.21.0",
+    "eslint-config-prettier": "10.0.1",
+    "eslint-plugin-prettier": "5.2.3",
     "eslint-plugin-vue": "9.32.0",
-    "fs-extra": "11.2.0",
-    "less": "4.2.1",
-    "npm-check-updates": "17.1.13",
-    "pkg-types": "1.3.0",
-    "postcss": "8.4.49",
-    "postcss-html": "1.7.0",
+    "fs-extra": "11.3.0",
+    "globals": "16.0.0",
+    "less": "4.2.2",
+    "npm-check-updates": "17.1.15",
+    "pkg-types": "1.3.1",
+    "postcss": "8.5.3",
+    "postcss-html": "1.8.0",
     "postcss-less": "6.0.0",
-    "prettier": "3.4.2",
-    "prettier-plugin-packagejson": "2.5.6",
+    "prettier": "3.5.2",
+    "prettier-plugin-packagejson": "2.5.8",
     "rimraf": "6.0.1",
-    "stylelint": "16.12.0",
+    "stylelint": "16.14.1",
+    "stylelint-config-recommended": "^15.0.0",
     "stylelint-config-recommended-less": "3.0.1",
-    "stylelint-config-recommended-vue": "1.5.0",
-    "stylelint-config-standard": "36.0.1",
-    "stylelint-config-standard-less": "3.0.1",
-    "stylelint-order": "6.0.4",
-    "stylelint-prettier": "5.0.2",
+    "stylelint-config-recommended-vue": "1.6.0",
+    "stylelint-config-standard": "37.0.0",
+    "stylelint-config-standard-less": "^3.0.1",
+    "stylelint-less": "^3.0.1",
+    "stylelint-prettier": "5.0.3",
     "ts-node": "10.9.2",
-    "turbo": "2.3.3",
-    "typescript": "5.7.2",
-    "unocss": "0.65.3",
-    "vite": "6.0.7"
+    "turbo": "2.4.2",
+    "typescript": "5.7.3",
+    "unocss": "66.0.0",
+    "vite": "6.1.1"
   },
   "keywords": [
     "typescript",
