@@ -1,4 +1,9 @@
-module.exports = {
+/**
+ * Copyright (c) 2013-Now http://jeesite.com All rights reserved.
+ * No deletion without permission, or be held responsible to law.
+ * @author ThinkGem
+ */
+export default {
   printWidth: 100,
   tabWidth: 2,
   useTabs: false,
@@ -12,7 +17,7 @@ module.exports = {
   arrowParens: 'always',
   insertPragma: false,
   requirePragma: false,
-  proseWrap: 'never',
+  proseWrap: 'preserve',
   htmlWhitespaceSensitivity: 'strict',
   endOfLine: 'auto',
   plugins: ['prettier-plugin-packagejson'],
